run_ -> mdtest::run
