+++ crates/ty_test/Cargo.toml
@@ -27,7 +27,6 @@ ty_vendored = { workspace = true }
 
 anyhow = { workspace = true }
 camino = { workspace = true }
-colored = { workspace = true }
 dunce = { workspace = true }
 salsa = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
